chore(suite): rewrite tokens table to table component
