+++ packages/components/src/components/Table/Table.tsx
@@ -1,48 +1,24 @@
 import { ReactNode } from 'react';
-import styled from 'styled-components';
-
-import { borders, Elevation, mapElevationToBackground } from '@trezor/theme';
 
 import { TableHeader } from './TableHeader';
 import { TableCell, TableCellProps } from './TableCell';
 import { TableRow } from './TableRow';
-import { useElevation } from '../ElevationContext/ElevationContext';
-import {
-    FrameProps,
-    FramePropsKeys,
-    pickAndPrepareFrameProps,
-    withFrameProps,
-} from '../../utils/frameProps';
-import { TransientProps } from '../../utils/transientProps';
+import { FrameProps, FramePropsKeys } from '../../utils/frameProps';
+import { TableBody } from './TableBody';
+import { Card } from '../Card/Card';
 
 export const allowedTableFrameProps = ['margin'] as const satisfies FramePropsKeys[];
 type AllowedFrameProps = Pick<FrameProps, (typeof allowedTableFrameProps)[number]>;
 
-const Container = styled.table<{ $elevation: Elevation } & TransientProps<AllowedFrameProps>>`
-    display: flex;
-    width: 100%;
-    flex-direction: column;
-    background: ${mapElevationToBackground};
-    border-radius: ${borders.radii.md};
-    box-shadow: ${({ theme, $elevation }) => $elevation === 1 && theme.boxShadowBase};
-
-    ${withFrameProps}
-`;
-
-interface TableProps {
+type TableProps = {
     children: ReactNode;
-}
+} & AllowedFrameProps;
 
-export const Table = ({ children, ...rest }: TableProps) => {
-    const frameProps = pickAndPrepareFrameProps(rest, allowedTableFrameProps);
-    const { elevation } = useElevation();
-
-    return (
-        <Container $elevation={elevation} {...frameProps}>
-            {children}
-        </Container>
-    );
-};
+export const Table = ({ children, margin }: TableProps) => (
+    <Card paddingType="none" margin={margin}>
+        <table>{children}</table>
+    </Card>
+);
 
 Table.Row = TableRow;
 Table.Cell = TableCell;
@@ -52,5 +28,6 @@ Table.HeaderCell = (props: TableCellProps) => (
         {props.children}
     </TableCell>
 );
+Table.Body = TableBody;
 
 export default Table;
