+++ packages/components/src/components/Table/TableBody.tsx
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react';
+
+export interface TableBodyProps {
+    children: ReactNode;
+}
+
+export const TableBody = ({ children }: TableBodyProps) => {
+    return <tbody>{children}</tbody>;
+};
